RPI/GUI/Script/requirements.txt
PyQt5==5.15.6
PyQt5-Qt5==5.15.2
PyQt5-sip==12.9.0
PyQtWebEngine== 5.15.5
PyQtWebEngine-Qt5==5.15.2
pyserial==3.5
folium==0.12.1.post1
pathlib==1.0.1
opencv-python==4.5.4.60
requests==2.27.1
urllib3==1.26.8
